Ensure battery saver preference is truly disabled while plugged in
The battery saver switch is disabled while the device is plugged in but
the preference is not. This seems like an oversight on Google's part. The
QS battery tile is disabled/grayed out so this preference should as well.

On top of that, it doesn't make sense to have this preference available to
the user if the device is plugged in. These options are useless unless you
unplug your device.

Change-Id: I22f69dc0468767b54423780f360fed9355a07b52

Signed-off-by: xyyx <xyyx@mail.ru>